requirements.txt
fastapi==0.116.1
openai==1.99.6
python-dotenv==1.0.1
uvicorn[standard]==0.32.1